+++ src/main/java/org/cardanofoundation/reeve/indexer/model/response/LEIResponse.java
@@ -19,5 +19,6 @@ public class LEIResponse {
 
     private boolean identityVerified;
     private String lei;
+    private String txHash;
 
 }